Memisahkan fungsi tertentu dalam file baru
