+++ login_reg_prof.py
@@ -1,47 +1,6 @@
-import json
-import os
-import time
+import os, time
+from fungsi import muat_data, simpan_data, validasi_tanggal_lahir, data_pengguna, tampilkan_profil
 
-FILE_DATABASE = "data_login.json"
-
-# Fungsi untuk memuat data dari file JSON
-def muat_data():
-    if os.path.exists(FILE_DATABASE):
-        with open(FILE_DATABASE, "r") as file:
-            return json.load(file)
-    #return {"users": {}, "admin": {"username": "admin", "password": "admin123"}}
-
-# Fungsi untuk menyimpan data ke file JSON
-def simpan_data(data):
-    with open(FILE_DATABASE, "w") as file:
-        json.dump(data, file, indent=4)
-
-def validasi_tanggal_lahir(tanggal):
-    try:
-        if not isinstance(tanggal, str):
-            return False
-        if len(tanggal) != 10 or tanggal[2] != "-" or tanggal[5] != "-":
-            return False
-        hari, bulan, tahun = tanggal.split("-")
-        if not (hari.isdigit() and bulan.isdigit() and tahun.isdigit()):
-            return False
-        hari, bulan, tahun = int(hari), int(bulan), int(tahun)
-        if not (1 <= hari <= 31 and 1 <= bulan <= 12 and len(str(tahun)) == 4):
-            return False
-        if bulan in [4, 6, 9, 11] and hari > 30:
-            return False
-        if bulan == 2:
-            if (tahun % 4 == 0 and (tahun % 100 != 0 or tahun % 400 == 0)):
-                if hari > 29:
-                    return False
-            else:
-                if hari > 28:
-                    return False
-        return True
-    except:
-        return False
-
-data_pengguna = muat_data()
 
 # Fungsi untuk menambahkan profil pengguna baru
 def register():
@@ -193,16 +152,6 @@ def register():
     time.sleep(2)
     os.system("cls")
     login_pengguna()
-
-# Fungsi untuk menampilkan data profil pengguna
-def tampilkan_profil(profil):
-    print("=== Profil Pengguna ===")
-    print(f"Nama: {profil['nama']}")
-    print(f"NIM: {profil['NIM']}")
-    print(f"Kelas: {profil['kelas']}")
-    print(f"Tanggal Lahir: {profil['tanggal_lahir']}")
-    print(f"No Telepon: {profil['no_telepon']}")
-    print(f"Email: {profil['email']}")
 
 # Fungsi login admin
 def login_admin():
